Moved loggedIn user permission fetching to Impl layer from Rest layer
