+++ components/apimgt/org.wso2.carbon.apimgt.core/src/main/java/org/wso2/carbon/apimgt/core/impl/APIPublisherImpl.java
@@ -195,6 +195,30 @@ public class APIPublisherImpl extends AbstractAPIManager implements APIPublisher
         return null;
     }
 
+    @Override
+    public API getAPIbyUUID(String uuid) throws APIManagementException {
+        API api = null;
+        try {
+            api = super.getAPIbyUUID(uuid);
+            if (api != null) {
+                api.setUserSpecificApiPermissions(getAPIPermissionsOfLoggedInUser(getUsername(), api));
+                String permissionString = api.getApiPermission();
+                if (!StringUtils.isEmpty(permissionString)) {
+                    api.setApiPermission(replaceGroupIdWithName(permissionString));
+                }
+            }
+        } catch (APIMgtDAOException e) {
+            String errorMsg = "Error occurred while retrieving API with id " + uuid;
+            log.error(errorMsg, e);
+            throw new APIManagementException(errorMsg, e, ExceptionCodes.APIMGT_DAO_EXCEPTION);
+        } catch (ParseException e) {
+            String errorMsg = "Error occurred while parsing the permission json string for API " + api.getName();
+            log.error(errorMsg, e);
+            throw new APIManagementException(errorMsg, e, ExceptionCodes.JSON_PARSE_ERROR);
+        }
+        return api;
+    }
+
     /**
      * Adds a new API to the system
      *
@@ -628,6 +652,114 @@ public class APIPublisherImpl extends AbstractAPIManager implements APIPublisher
             String errorMessage = "There are invalid roles in the permission string";
             log.error(errorMessage, e);
             throw new APIManagementException(errorMessage, e, ExceptionCodes.UNSUPPORTED_ROLE);
+        }
+        return updatedPermissionArray.toJSONString();
+    }
+
+    /**
+     * This method retrieves the set of overall permissions for a given api for the logged in user
+     *
+     * @param loggedInUserName - Logged in user
+     * @param api - The API whose permissions for the logged in user is retrieved
+     * @return The overall list of permissions for the given API for the logged in user
+     */
+    private List<String> getAPIPermissionsOfLoggedInUser(String loggedInUserName, API api)
+            throws APIManagementException {
+        Set<String> permissionArrayForUser = new HashSet();
+        //Setting default permissions for API
+        permissionArrayForUser.add(APIMgtConstants.Permission.READ);
+        permissionArrayForUser.add(APIMgtConstants.Permission.UPDATE);
+        permissionArrayForUser.add(APIMgtConstants.Permission.DELETE);
+        Map<String, Integer> permissionMap = api.getPermissionMap();
+
+        //TODO: Remove the check for admin after IS adds an ID to admin user
+        if (permissionMap != null && !permissionMap.isEmpty() && !"admin".equals(loggedInUserName)) {
+            String userId = getIdentityProvider().getIdOfUser(loggedInUserName);
+            List<String> loggedInUserRoles = getIdentityProvider().getRoleIdsOfUser(userId);
+            List<String> permissionRoleList = getRolesFromPermissionMap(permissionMap);
+            List<String> rolesOfUserWithAPIPermissions = null;
+            //To prevent a possible null pointer exception
+            if (loggedInUserRoles == null) {
+                loggedInUserRoles = new ArrayList<>();
+            }
+            //get the intersection - retainAll() transforms first set to the result of intersection
+            loggedInUserRoles.retainAll(permissionRoleList);
+            if (!loggedInUserRoles.isEmpty()) {
+                rolesOfUserWithAPIPermissions = loggedInUserRoles;
+            }
+            if (rolesOfUserWithAPIPermissions != null) {
+                //remove all elements from set
+                permissionArrayForUser.clear();
+                for (String role : rolesOfUserWithAPIPermissions) {
+                    Integer permission = permissionMap.get(role);
+                    if (permission == APIMgtConstants.Permission.READ_PERMISSION) {
+                        permissionArrayForUser.add(APIMgtConstants.Permission.READ);
+                    } else if (permission == (APIMgtConstants.Permission.READ_PERMISSION
+                            + APIMgtConstants.Permission.UPDATE_PERMISSION)) {
+                        permissionArrayForUser.add(APIMgtConstants.Permission.READ);
+                        permissionArrayForUser.add(APIMgtConstants.Permission.UPDATE);
+                    } else if (permission == (APIMgtConstants.Permission.READ_PERMISSION
+                            + APIMgtConstants.Permission.DELETE_PERMISSION)) {
+                        permissionArrayForUser.add(APIMgtConstants.Permission.READ);
+                        permissionArrayForUser.add(APIMgtConstants.Permission.DELETE);
+                    } else if (permission
+                          == APIMgtConstants.Permission.READ_PERMISSION + APIMgtConstants.Permission.UPDATE_PERMISSION
+                            + APIMgtConstants.Permission.DELETE_PERMISSION) {
+                        permissionArrayForUser.add(APIMgtConstants.Permission.READ);
+                        permissionArrayForUser.add(APIMgtConstants.Permission.UPDATE);
+                        permissionArrayForUser.add(APIMgtConstants.Permission.DELETE);
+                    }
+                }
+            }
+        }
+        List<String> finalAggregatedPermissionList = new ArrayList<>();
+        finalAggregatedPermissionList.addAll(permissionArrayForUser);
+        return finalAggregatedPermissionList;
+    }
+
+    /**
+     * This method is used to extract the groupIds or roles from the permissionMap
+     *
+     * @param permissionMap - The map containing the group IDs(roles) and their permissions
+     * @return - The list of groupIds specified for permissions
+     */
+    private List<String> getRolesFromPermissionMap(Map<String, Integer> permissionMap) {
+        List<String> permissionRoleList = new ArrayList<>();
+        for (String groupId : permissionMap.keySet()) {
+            permissionRoleList.add(groupId);
+        }
+        return permissionRoleList;
+    }
+
+    /**
+     * This method replaces the groupId field's value of the api permisisons string to the role name before sending to
+     * frontend
+     *
+     * @param permissionString - permissions string containing role ids in the groupId field
+     * @return the permission string replacing the groupId field's value to role name
+     * @throws ParseException - if there is an erro rparsing the permission json
+     * @throws APIManagementException - if there is an error getting the IdentityProvider instance
+     */
+    private String replaceGroupIdWithName(String permissionString) throws ParseException, APIManagementException {
+        JSONArray updatedPermissionArray = new JSONArray();
+        JSONParser jsonParser = new JSONParser();
+        JSONArray originalPermissionArray = (JSONArray) jsonParser.parse(permissionString);
+
+        for (Object permissionObj : originalPermissionArray) {
+            JSONObject jsonObject = (JSONObject) permissionObj;
+            String groupId = (String) jsonObject.get(APIMgtConstants.Permission.GROUP_ID);
+            try {
+                String groupName = getIdentityProvider().getRoleName(groupId);
+                JSONObject updatedPermissionJsonObj = new JSONObject();
+                updatedPermissionJsonObj.put(APIMgtConstants.Permission.GROUP_ID, groupName);
+                updatedPermissionJsonObj.put(APIMgtConstants.Permission.PERMISSION,
+                        jsonObject.get(APIMgtConstants.Permission.PERMISSION));
+                updatedPermissionArray.add(updatedPermissionJsonObj);
+            } catch (IdentityProviderException e) {
+                //lets the execution continue after logging the exception
+                String errorMessage = "Role with ID " + groupId + " no longer exists in the system";
+                log.error(errorMessage, e);
+            }
         }
         return updatedPermissionArray.toJSONString();
     }
