+++ components/apimgt/org.wso2.carbon.apimgt.rest.api.publisher/src/gen/java/org/wso2/carbon/apimgt/rest/api/publisher/dto/APIInfoDTO.java
@@ -5,8 +5,6 @@ import com.fasterxml.jackson.annotation.JsonProperty;
 import com.fasterxml.jackson.annotation.JsonCreator;
 import io.swagger.annotations.ApiModel;
 import io.swagger.annotations.ApiModelProperty;
-import java.util.ArrayList;
-import java.util.List;
 import java.util.Objects;
 
 /**
@@ -36,9 +34,6 @@ public class APIInfoDTO   {
 
   @JsonProperty("workflowStatus")
   private String workflowStatus = null;
-
-  @JsonProperty("userPermissionsForApi")
-  private List<String> userPermissionsForApi = new ArrayList<String>();
 
   public APIInfoDTO id(String id) {
     this.id = id;
@@ -184,29 +179,6 @@ public class APIInfoDTO   {
     this.workflowStatus = workflowStatus;
   }
 
-  public APIInfoDTO userPermissionsForApi(List<String> userPermissionsForApi) {
-    this.userPermissionsForApi = userPermissionsForApi;
-    return this;
-  }
-
-  public APIInfoDTO addUserPermissionsForApiItem(String userPermissionsForApiItem) {
-    this.userPermissionsForApi.add(userPermissionsForApiItem);
-    return this;
-  }
-
-   /**
-   * LoggedIn user permissions for the API 
-   * @return userPermissionsForApi
-  **/
-  @ApiModelProperty(example = "[&quot;READ&quot;,&quot;UPDATE&quot;]", value = "LoggedIn user permissions for the API ")
-  public List<String> getUserPermissionsForApi() {
-    return userPermissionsForApi;
-  }
-
-  public void setUserPermissionsForApi(List<String> userPermissionsForApi) {
-    this.userPermissionsForApi = userPermissionsForApi;
-  }
-
 
   @Override
   public boolean equals(java.lang.Object o) {
@@ -224,13 +196,12 @@ public class APIInfoDTO   {
         Objects.equals(this.version, apIInfo.version) &&
         Objects.equals(this.provider, apIInfo.provider) &&
         Objects.equals(this.lifeCycleStatus, apIInfo.lifeCycleStatus) &&
-        Objects.equals(this.workflowStatus, apIInfo.workflowStatus) &&
-        Objects.equals(this.userPermissionsForApi, apIInfo.userPermissionsForApi);
+        Objects.equals(this.workflowStatus, apIInfo.workflowStatus);
   }
 
   @Override
   public int hashCode() {
-    return Objects.hash(id, name, description, context, version, provider, lifeCycleStatus, workflowStatus, userPermissionsForApi);
+    return Objects.hash(id, name, description, context, version, provider, lifeCycleStatus, workflowStatus);
   }
 
   @Override
@@ -246,7 +217,6 @@ public class APIInfoDTO   {
     sb.append("    provider: ").append(toIndentedString(provider)).append("\n");
     sb.append("    lifeCycleStatus: ").append(toIndentedString(lifeCycleStatus)).append("\n");
     sb.append("    workflowStatus: ").append(toIndentedString(workflowStatus)).append("\n");
-    sb.append("    userPermissionsForApi: ").append(toIndentedString(userPermissionsForApi)).append("\n");
     sb.append("}");
     return sb.toString();
   }
