+++ components/web/components/org.wso2.carbon.apimgt.publisher.commons.ui/src/main/fragments/api-details/api-details.hbs
@@ -51,10 +51,10 @@
                         </li>
                     {{/if}}
                     {{#if @params.publisher_app}}
-                    <li id="tab-6" role="presentation"><a href="#actrl-tab" role="tab" aria-controls="actrl-tab"
-                                                          data-toggle="tab"><i
-                            class="fw fw-contract"></i>&nbsp; {{i18n "Access Control"}}</a>
-                    </li>
+                        <li id="tab-6" role="presentation"><a href="#actrl-tab" role="tab" aria-controls="actrl-tab"
+                                                              data-toggle="tab"><i
+                                class="fw fw-contract"></i>&nbsp; {{i18n "Access Control"}}</a>
+                        </li>
                     {{/if}}
                     <li id="tab-7" role="presentation"><a href="#mediation-tab" role="tab" aria-controls="mediation-tab"
                                                           data-toggle="tab"><i
